radius=input() area=3.14*int(radius)*int(radius) print('%.2f' %area)

+++ Area_of_a_circle.py
@@ -1,3 +1,3 @@
 radius=input()
-area=3.14*float(radius)*float(radius)
+area=3.14*int(radius)*int(radius)
 print('%.2f' %area)
